Used react-router to separate components. Tables mockup displays correctly again.
