+++ src/App.css
@@ -15,6 +15,10 @@
 	height: inherit;
 }
 
+#AppHelp {
+	height: 100%;
+}
+
 #AppMenuItemHexpop {
 	margin-top: 1rem;
 }
