+++ src/Help.css
@@ -1,0 +1,8 @@
+#Help {
+	height: 100%;
+}
+
+#HelpSegment {
+	height: 100%;
+	overflow: auto;
+}
