+++ src/Project.css
@@ -1,0 +1,17 @@
+#Workspace {
+	height: inherit;
+}
+
+#WorkspaceSegmentScroll {
+	height: calc(100% - 2.4rem);
+	overflow: auto;
+}
+
+#WorkspaceControls {
+	height: inherit;
+}
+
+#WorkspaceControlsSegmentScroll {
+	height: calc(100% - 7.15rem);
+	overflow: auto;
+}
