+++ src/Project/Tables.css
@@ -1,0 +1,7 @@
+#TablesWorkspace {
+	height: inherit;
+}
+
+#TablesHelp {
+	height: 100%;
+}
